Resolved issue in importing files.
Former-commit-id: 84bb0718e7ea990e2f96f49268affb9cb04cf3b6

+++ src/com/vimukti/accounter/web/server/FinanceTool.java
@@ -4885,7 +4885,9 @@ public class FinanceTool {
 					}
 				}
 			}
-			file.delete();
+			if (exceptions.isEmpty()) {
+				file.delete();
+			}
 			exceptions.put(0, successCount);
 		} catch (Exception e) {
 			e.printStackTrace();
